Glitchy prototypal inheritance (for now)
I'm defining all members, etc. in the CTORs of the Parent and
Child classes since I got nothing but errors when using
X.prototype = { obj literal }. I need a clean codebase to play
with this some...

Also, there's some random error that cropped up since before I
started to work on this:

	elem.style is undefined
		display = elem.style.display;

	jquery.js (line 7440)

Not sure what is causing it, but it doesn't seem to cause problems.